change page transitions, navigation to router
