+++ app/lib/components/parts/competition_button.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:go_router/go_router.dart';
 
 class CompetitionButton extends ConsumerWidget {
 
@@ -11,9 +12,7 @@ class CompetitionButton extends ConsumerWidget {
         clipBehavior: Clip.hardEdge,
         child: InkWell(
           splashColor: Colors.blue.withAlpha(60),
-          onTap: () {
-            Navigator.pushNamed(context, '/Entry');
-          },
+          onTap: () => context.push('/entry'),
           child: const SizedBox(
             width: 300,
             height: 100,
