+++ app/lib/components/parts/server_button.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:go_router/go_router.dart';
 
 class ServerButton extends ConsumerWidget {
   @override
@@ -10,9 +11,7 @@ class ServerButton extends ConsumerWidget {
         clipBehavior: Clip.hardEdge,
         child: InkWell(
           splashColor: Colors.blue.withAlpha(60),
-          onTap: () {
-            Navigator.pushNamed(context, '/Server');
-          },
+          onTap: () => context.push('/server'),
           child: const SizedBox(
             width: 300,
             height: 100,
