+++ app/lib/main.dart
@@ -2,10 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import 'components/pages/score_server_page.dart';
-import 'components/pages/competition_entry_page.dart';
-import 'components/parts/server_button.dart';
-import 'components/parts/competition_button.dart';
+import 'router.dart' as Router;
 
 Future main() async {
   runApp(
@@ -20,36 +17,12 @@ class TetrisApp extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    return MaterialApp(
+    return MaterialApp.router(
       title: 'Tetris HP',
       theme: ThemeData(
         primarySwatch: Colors.green,
       ),
-      routes: {
-        '/': (context) => EntryPage(),
-        '/Server': (context) => ScoreServerPage(),
-        '/Entry': (context) => CompetitionEntryPage()
-      }
-    );
-  }
-}
-
-class EntryPage extends ConsumerWidget {
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    return Scaffold(
-      appBar: AppBar(title: const Text('Tetris HP')),
-      body: Container(
-        child: ListView(
-          scrollDirection: Axis.horizontal,
-          padding: const EdgeInsets.all(8),
-          children: <Widget>[
-            CompetitionButton(),
-            ServerButton()
-          ],
-        ),
-      )
+      routerConfig: Router.Router,
     );
   }
 }
