+++ app/pubspec.yaml
@@ -44,6 +44,7 @@ dependencies:
   freezed_annotation: ^2.2.0
   json_annotation: ^4.7.0
   pluto_grid: ^5.4.9
+  go_router: ^6.2.0
 
 dev_dependencies:
   flutter_test:
